Implement Sentiment Analysis System: Compare NLTK, spaCy, and ML Models for High-Accuracy Review Classification
This study focuses on creating a system to automatically detect negative movie reviews. The goal is to train a model that can classify reviews as either positive or negative, aiming for an F1 score of at least 0.85 to ensure high accuracy.

To achieve this, we compare three different approaches: using NLTK with TF-IDF and Linear Regression, spaCy with TF-IDF and Linear Regression, and spaCy with TF-IDF and LGBMClassifier. Through this comparison, we aim to identify the most effective method for sentiment analysis in movie reviews, balancing both performance and efficiency.